portfolio section redesign. filters implemented
